+++ Models/Data/RankUser.cs
@@ -4,7 +4,7 @@ namespace DotnetCoreServer.Models
 {
     public class RankUser
     {
-        public long UserID;
+        public int UserID;
         public string FacebookID;
         public string FacebookName;
         public string FacebookPhotoURL;
